refactor: TextView에 AtrtributeString 적용 방식 변경
UITextView Extension 제거하고 String Extension을 통해 AttributeString 적용
